ontix/ml_service/requirements.txt
# HDBSCAN + UMAP
hdbscan>=0.8.38
umap-learn>=0.5.5

# Dependencies
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<1.6.0

# HTTP API
fastapi>=0.109.0
uvicorn>=0.27.0

# OpenAI (for cluster naming)
openai>=1.3.0

# Experiment tools
psycopg2-binary>=2.9.0
tabulate>=0.9.0
